Add more bindings for the JoinParam4WithoutLogin  and fixed memory reference bugs
Added new parameters for audio raw data handling in JoinParam4WithoutLogin. In withoutloginuserJoin, change vanityID, customer_key, userZak to have the same treatment as meetingNumber and psw, the code was allocating a c_str() in the stack and that corrupted the value stored in the struct; pointing to an overrwritten value in the struct. This prevented using those values for joining a meeting. Now, nanobind should correctly track the reference counters.

This was tested with the library version 6.3.5 (https://developers.zoom.us/changelog/meeting-sdk/linux/6.3.5/); I think this is the version used in this library!